refactor: move all view commands into one file
